[REFACTOR] page size change to respond with total number of elements

+++ src/main/java/com/gabojait/gabojaitspring/common/dto/DefaultMultiResDto.java
@@ -13,7 +13,7 @@ public class DefaultMultiResDto<T> extends DefaultNoResDto {
     private PageDataDto responseData;
 
     @Builder(builderMethodName = "multiDataBuilder", builderClassName = "multiDataBuilder")
-    public DefaultMultiResDto(String responseCode, String responseMessage, final T data, final int size) {
+    public DefaultMultiResDto(String responseCode, String responseMessage, final T data, final long size) {
         super(responseCode, responseMessage);
 
         this.responseData = PageDataDto.builder()
